Rough draft of new time-series animation structure
Now the big data series remain static within their own svg,
and and an outer svg has a panning viewPort.
The inner svg is still important because it defines the height
of the datapoints relatively. This allows me to define everything
in terms of percents, and if I hypothetically want to change the
height of one graph, I can do so by changing the height of both the
inner and outer svgs, and the circles won't be squashed or anything :)